requirements.txt
asgiref==3.8.1
Django==4.2
django-bootstrap5==25.1
django-filter==25.1
djangorestframework==3.15.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
